@@ -1,6 +1,6 @@
 {
   "name": "formsy-react",
-  "version": "0.17.0",
+  "version": "0.19.5",
   "description": "A form input builder and validator for React JS",
   "repository": {
     "type": "git",
@@ -8,8 +8,7 @@
   },
   "main": "lib/main.js",
   "scripts": {
-    "start": "webpack-dev-server --content-base build",
-    "deploy": "NODE_ENV=production webpack -p --config webpack.production.config.js",
+    "build": "NODE_ENV=production webpack -p --config webpack.production.config.js",
     "test": "babel-node testrunner",
     "examples": "webpack-dev-server --config examples/webpack.config.js --content-base examples",
     "prepublish": "babel ./src/ -d ./lib/"
@@ -24,24 +23,27 @@
     "react-component"
   ],
   "dependencies": {
-    "form-data-to-object": "^0.1.0"
+    "form-data-to-object": "^0.2.0"
   },
   "devDependencies": {
-    "babel": "^5.6.4",
-    "babel-core": "^5.1.11",
-    "babel-loader": "^5.0.0",
+    "babel-cli": "^6.6.5",
+    "babel-loader": "^6.2.4",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-react": "^6.5.0",
+    "babel-preset-stage-2": "^6.5.0",
+    "create-react-class": "^15.6.0",
     "jsdom": "^6.5.1",
-    "lolex": "^1.3.2",
     "nodeunit": "^0.9.1",
-    "react": "^0.14.0-rc1",
-    "react-addons-pure-render-mixin": "^0.14.2",
-    "react-addons-test-utils": "^0.14.0-rc1",
-    "react-dom": "^0.14.0-rc1",
-    "sinon": "^1.17.1",
-    "webpack": "^1.7.3",
-    "webpack-dev-server": "^1.7.0"
+    "prop-types": "^15.5.10",
+    "react": "^15.0.0",
+    "react-addons-pure-render-mixin": "^15.0.0",
+    "react-addons-test-utils": "^15.0.0",
+    "react-dom": "^15.0.0",
+    "sinon": "^1.17.3",
+    "webpack": "^1.12.14",
+    "webpack-dev-server": "^1.14.1"
   },
   "peerDependencies": {
-    "react": "^0.14.0"
+    "react": "^0.14.0 || ^15.0.0"
   }
 }
