@@ -20,8 +20,8 @@ const App = React.createClass({
   render() {
     return (
       <Form onSubmit={this.submit} onValid={this.enableButton} onInvalid={this.disableButton} className="login">
-        <MyInput name="email" title="Email" validations="isEmail" validationError="This is not a valid email" required />
-        <MyInput name="password" title="Password" type="password" required />
+        <MyInput value="" name="email" title="Email" validations="isEmail" validationError="This is not a valid email" required />
+        <MyInput value="" name="password" title="Password" type="password" required />
         <button type="submit" disabled={!this.state.canSubmit}>Submit</button>
       </Form>
     );
