@@ -1,8 +1,9 @@
 var React = global.React || require('react');
+var createReactClass = require('create-react-class');
 var Mixin = require('./Mixin.js');
 module.exports = function () {
   return function (Component) {
-    return React.createClass({
+    return createReactClass({
       mixins: [Mixin],
       render: function () {
         return React.createElement(Component, {
